requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.10.0
    # via openai
keyboard==0.13.5
    # via push-to-talk (pyproject.toml)
openai==1.97.1
    # via push-to-talk (pyproject.toml)
pyaudio==0.2.14
    # via push-to-talk (pyproject.toml)
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via push-to-talk (pyproject.toml)
pyautogui==0.9.54
    # via push-to-talk (pyproject.toml)
pyperclip==1.8.2
    # via push-to-talk (pyproject.toml)
simpleaudio==1.0.4
    # via push-to-talk (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
